infra: home unzip in MudletArchive and the memory query in TBuffer
