+++ src/CMakeLists.txt
@@ -106,6 +106,7 @@ set(mudlet_SRCS
     MMCPClient.cpp
     MMCPServer.cpp
     mudlet.cpp
+    MudletArchive.cpp
     MudletInstanceCoordinator.cpp
     MxpTag.cpp
     OAuthClientFlow.cpp
@@ -359,6 +360,7 @@ set(mudlet_HDRS
     MMCPClient.h
     MMCPServer.h
     mudlet.h
+    MudletArchive.h
     MudletInstanceCoordinator.h
     MxpTag.h
     OAuthClientFlow.h
